common: Store just an array in `Ipv*Addrs` structs
Addresses represented by `0` or arrays with zeros are wildcard. So when the
first element of the array is `0` or `[0u8; 16]`, we can just assume
that we have a wildcard policy. Therefore we can get rid of the `all`
field.

`len` field would be useful to reduce the number of iterations when
matching, but it trips the verifier.